Move transform related properties from effects to box.
