+++ components/style/properties/properties.mako.rs
@@ -1336,7 +1336,7 @@ impl ComputedValues {
             return transform_style::T::flat;
         }
 
-        if effects.transform_style == transform_style::T::auto {
+        if box_.transform_style == transform_style::T::auto {
             if box_.transform.0.is_some() {
                 return transform_style::T::flat;
             }
@@ -1346,7 +1346,7 @@ impl ComputedValues {
         }
 
         // Return the computed value if not overridden by the above exceptions
-        effects.transform_style
+        box_.transform_style
     }
 
     pub fn transform_requires_layer(&self) -> bool {
