+++ components/style/servo/restyle_damage.rs
@@ -227,7 +227,7 @@ fn compute_damage(old: &ServoComputedValues, new: &ServoComputedValues) -> Servo
         get_position.top, get_position.left,
         get_position.right, get_position.bottom,
         get_effects.opacity,
-        get_box.transform, get_effects.transform_style, get_effects.transform_origin,
+        get_box.transform, get_box.transform_style, get_box.transform_origin,
         get_box.perspective, get_box.perspective_origin
     ]) || add_if_not_equal!(old, new, damage,
                             [REPAINT], [
